FEAT: New UI and added startTrack interface callback and clear all download btn
Changed Ui, from all progress bar to only the first downloading track. With this the load on the recyclerView is reduced, since all those values are static now.

Also changed Ui to show if no download in their. All added a fixed downloading item, so only the info of the downloading track is displayed.

The Ui of the fixed item is changed with the startTrackDownload callback.

Removed the resource heavy adapter change calls.

added btn to clear all downloading files.