FEAT: Migrated to Ind level Threads for Data Update
Added new UpdateDatabaseService that can update data even in the background.
This will help to update the track data with server call.
Also if the user quit the app after opening still the data will be updated.
Also the check for local data was changed to artist track count based system.